# app/docs/swagger.yaml
definitions:
  controller.CreateTaskRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  model.AppFile:
    properties:
      createdAt:
        type: string
      fileType:
        type: string
      filename:
        type: string
      id:
        type: integer
      taskID:
        description: Foreign key
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
  model.Task:
    properties:
      completed:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/model.AppFile'
        type: array
      mesh:
        $ref: '#/definitions/model.AppFile'
      status:
        $ref: '#/definitions/model.TaskStatus'
      title:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  model.TaskStatus:
    enum:
    - SUCCESS
    - INPROGRESS
    - FAILED
    - INITIAL
    type: string
    x-enum-varnames:
    - SUCCESS
    - INPROGRESS
    - FAILED
    - INITIAL
info:
  contact: {}
paths:
  /tasks:
    post:
      consumes:
      - application/json
      description: Creates a new task for the authenticated user
      parameters:
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /tasks/{taskID}/start:
    post:
      consumes:
      - application/json
      description: Uploads files to a task
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateTaskRequest'
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload files to a task
      tags:
      - tasks
  /tasks/{taskID}/upload:
    post:
      consumes:
      - application/json
      description: Uploads files to a task
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      - description: Files to upload
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload files to a task
      tags:
      - tasks
swagger: "2.0"